==> scripts/forecasting_models/cross_validation/01_ml_prob_sktime_cv.ipynb <==
pip install shap

# Libraries
# ==============================================================================
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import shap
from sklearn.inspection import PartialDependenceDisplay

from skforecast.plot import set_dark_theme
from skforecast.preprocessing import series_long_to_dict
from skforecast.preprocessing import exog_long_to_dict
from skforecast.ForecasterAutoregMultiSeries import ForecasterAutoregMultiSeries
from skforecast.model_selection_multiseries import backtesting_forecaster_multiseries
from skforecast.model_selection_multiseries import bayesian_search_forecaster_multiseries
from sklearn.metrics import mean_squared_error
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import FunctionTransformer
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import make_pipeline
from category_encoders.ordinal import OrdinalEncoder


from lightgbm import LGBMRegressor
from sklearn.ensemble import RandomForestRegressor
import lightgbm as lgb
from xgboost import XGBRegressor
from sklearn.linear_model import LinearRegression

from sklearn import metrics
from sklearn.metrics import  r2_score

import statistics
from  datetime import datetime, timedelta

# read data

stock_df = pd.read_csv('D:/PhD_project/data/tidy_data/stock_tsb_zero_remove.csv')

del stock_df['Unnamed: 0']

stock_df.head()

# make the yearmonth column as date format

stock_df['yearmonth'] = pd.to_datetime(stock_df['yearmonth']) + pd.offsets.MonthEnd(0)

stock_df = stock_df.sort_values(by = 'yearmonth')

stock_df.head()

# feature engineering

# add rolling statistics

# dummy function to count zeros
count_func = lambda x: (x==0).sum()

stock_df['year'] = pd.DatetimeIndex(stock_df['yearmonth']).year
stock_df['month'] = pd.DatetimeIndex(stock_df['yearmonth']).month

number_lags = 4

for lag in range(1, number_lags + 1):
    stock_df['lag_' + str(lag)] = stock_df.groupby(['site_code', 'product_code'])['stock_distributed'].shift(lag).fillna(0)
    

stock_df['lag_rolling_mean_1234'] = stock_df[['lag_1', 'lag_2', 'lag_3', 'lag_4']].mean(axis=1)

stock_df['rolling_max_4'] = stock_df.groupby('id')['stock_distributed'].rolling(4).max().reset_index(level = 0, drop = True).fillna(0)

stock_df['rolling_zero_per_4'] = (stock_df.groupby('id')['stock_distributed'].rolling(4).apply(count_func)/4).reset_index(level = 0, drop = True).fillna(0)

stock_tidy = stock_df.drop(['lag_1', 'lag_2', 'lag_3', 'lag_4', 'unique_id'], axis = 1)

stock_tidy.head(5)

# categorical encoding

enc = OrdinalEncoder()

stock_tidy[['region', 'district', 'site_code', 'site_type', 'product_type', 'product_code']] = enc.fit_transform(stock_tidy[['region', 'district', 'site_code', 'site_type', 'product_type', 'product_code']])


# create series and exogenous data

series = stock_tidy[['yearmonth', 'id', 'stock_distributed']]

exog = stock_tidy[['id', 'yearmonth', 'region', 'district', 'site_code', 'site_type', 'product_type', 'product_code', 'year', 'month', 'lag_rolling_mean_1234', 'rolling_max_4', 'rolling_zero_per_4']]

# Transform series and exog to dictionaries

series_dict = series_long_to_dict(
    data      = series,
    series_id = 'id',
    index     = 'yearmonth',
    values    = 'stock_distributed',
    freq      = 'M'
)

exog_dict = exog_long_to_dict(
    data      = exog,
    series_id = 'id',
    index     = 'yearmonth',
    freq      = 'M'
)

# Partition data in train and test - Cross validation

end_train_set1 = '2019-07-31'
start_test_set1 = '2019-08-31'
end_test_set1 = '2019-10-31'

series_dict_train_set1 = {k: v.loc[:end_train_set1,] for k, v in series_dict.items()}
exog_dict_train_set1   = {k: v.loc[:end_train_set1,] for k, v in exog_dict.items()}
series_dict_test_set1  = {k: v.loc[start_test_set1:end_test_set1,] for k, v in series_dict.items()}
exog_dict_test_set1    = {k: v.loc[start_test_set1:end_test_set1,] for k, v in exog_dict.items()}


end_train_set2 = '2019-08-31'
start_test_set2 = '2019-09-30'
end_test_set2 = '2019-11-30'

series_dict_train_set2 = {k: v.loc[:end_train_set2,] for k, v in series_dict.items()}
exog_dict_train_set2   = {k: v.loc[:end_train_set2,] for k, v in exog_dict.items()}
series_dict_test_set2  = {k: v.loc[start_test_set2:end_test_set2,] for k, v in series_dict.items()}
exog_dict_test_set2    = {k: v.loc[start_test_set2:end_test_set2,] for k, v in exog_dict.items()}


end_train_set3 = '2019-09-30'
start_test_set3 = '2019-10-31'

series_dict_train_set3 = {k: v.loc[:end_train_set3,] for k, v in series_dict.items()}
exog_dict_train_set3   = {k: v.loc[:end_train_set3,] for k, v in exog_dict.items()}
series_dict_test_set3  = {k: v.loc[start_test_set3:,] for k, v in series_dict.items()}
exog_dict_test_set3    = {k: v.loc[start_test_set3:,] for k, v in exog_dict.items()}

# LGBM

# Fit forecaster - CV 1

regressor_lgbm = LGBMRegressor(
                boosting_type = 'gbdt',
                metric = 'mae',
                learning_rate = 0.1,
                num_iterations = 200,
                n_estimators = 100,
               #device_type = 'gpu',
                objective = 'poisson')
                #max_bins = 128)

forecaster_lgbm = ForecasterAutoregMultiSeries(
                 regressor          = regressor_lgbm, 
                 transformer_series = None,
                 lags               = 4,  
                 dropna_from_series = False
             )

forecaster_lgbm.fit(series=series_dict_train_set1, exog=exog_dict_train_set1, suppress_warnings=True)

forecaster_lgbm

# Feature importances
feat_lgbm = forecaster_lgbm.get_feature_importances()

# save file
feat_lgbm.to_csv('D:/PhD_project/data/results/lgbm_feat_imp_skt_cv1.csv', index = False)

# Predict - CV1

boot = 1000

%time predictions_lgbm_cv1 = forecaster_lgbm.predict_bootstrapping(steps=3, exog=exog_dict_test_set1, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_lgbm_cv1.keys())

# Create an empty DataFrame
df = pd.DataFrame(columns=['yearmonth', 'id', 'model', 'horizon'] + [f'X_{i}' for i in range(boot)])
df

for i in pred_id:
    
    df_test = predictions_lgbm_cv1[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'lgbm'
    
    df_test['horizon'] = 1

    df = pd.concat([df, df_test])
    

df

# update forecaster with new data
forecaster_lgbm.fit(series=series_dict_train_set2, exog=exog_dict_train_set2, suppress_warnings=True)

# Predict - CV2
%time predictions_lgbm_cv2 = forecaster_lgbm.predict_bootstrapping(steps=3, exog=exog_dict_test_set2, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_lgbm_cv2.keys())

for i in pred_id:
    
    df_test = predictions_lgbm_cv2[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'lgbm'
    
    df_test['horizon'] = 2

    df = pd.concat([df, df_test])

# Feature importances
feat_lgbm = forecaster_lgbm.get_feature_importances()

# save file
feat_lgbm.to_csv('D:/PhD_project/data/results/lgbm_feat_imp_skt_cv2.csv', index = False)

# update forecaster with new data
forecaster_lgbm.fit(series=series_dict_train_set3, exog=exog_dict_train_set3, suppress_warnings=True)

# Predict - CV2
%time predictions_lgbm_cv3 = forecaster_lgbm.predict_bootstrapping(steps=3, exog=exog_dict_test_set3, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_lgbm_cv3.keys())

for i in pred_id:
    
    df_test = predictions_lgbm_cv3[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'lgbm'
    
    df_test['horizon'] = 3

    df = pd.concat([df, df_test])

# Feature importances
feat_lgbm = forecaster_lgbm.get_feature_importances()

# save file
feat_lgbm.to_csv('D:/PhD_project/data/results/lgbm_feat_imp_skt_cv3.csv', index = False)

df

# save file

df.to_csv('D:/PhD_project/data/predictions/lgbm_forecasts_skt_cv.csv', index = False)

# XGB

# Fit forecaster

regressor_xgb = XGBRegressor(tree_method = 'hist',
                            objective = 'count:poisson',
                            enable_categorical = True)

forecaster_xgb = ForecasterAutoregMultiSeries(
                 regressor          = regressor_xgb, 
                 transformer_series = None,
                 lags               = 4,  
                 dropna_from_series = False
             )

forecaster_xgb.fit(series=series_dict_train_set1, exog=exog_dict_train_set1, suppress_warnings=True)

forecaster_xgb

# Feature importances
feat_xgb = forecaster_xgb.get_feature_importances()

# save file
feat_xgb.to_csv('D:/PhD_project/data/results/xgb_feat_imp_skt_cv1.csv', index = False)

# Predict - CV1

%time predictions_xgb_cv1 = forecaster_xgb.predict_bootstrapping(steps=3, exog=exog_dict_test_set1, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_xgb_cv1.keys())

# Create an empty DataFrame
df = pd.DataFrame(columns=['yearmonth', 'id', 'model', 'horizon'] + [f'X_{i}' for i in range(boot)])
df

for i in pred_id:
    
    df_test = predictions_xgb_cv1[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'xgb'
    
    df_test['horizon'] = 1

    df = pd.concat([df, df_test])

# update forecaster with new data
forecaster_xgb.fit(series=series_dict_train_set2, exog=exog_dict_train_set2, suppress_warnings=True)

# Predict - CV2
%time predictions_xgb_cv2 = forecaster_xgb.predict_bootstrapping(steps=3, exog=exog_dict_test_set2, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_xgb_cv2.keys())

for i in pred_id:
    
    df_test = predictions_xgb_cv2[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'xgb'
    
    df_test['horizon'] = 2

    df = pd.concat([df, df_test])

# Feature importances
feat_xgb = forecaster_xgb.get_feature_importances()

# save file
feat_xgb.to_csv('D:/PhD_project/data/results/xgb_feat_imp_skt_cv2.csv', index = False)

# update forecaster with new data
forecaster_xgb.fit(series=series_dict_train_set3, exog=exog_dict_train_set3, suppress_warnings=True)

# Predict - CV3
%time predictions_xgb_cv3 = forecaster_xgb.predict_bootstrapping(steps=3, exog=exog_dict_test_set3, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_xgb_cv3.keys())

for i in pred_id:
    
    df_test = predictions_xgb_cv3[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'xgb'
    
    df_test['horizon'] = 3

    df = pd.concat([df, df_test])

# Feature importances
feat_xgb = forecaster_xgb.get_feature_importances()

# save file
feat_xgb.to_csv('D:/PhD_project/data/results/xgb_feat_imp_skt_cv3.csv', index = False)

# save file

df.to_csv('D:/PhD_project/data/predictions/xgb_forecasts_skt_cv.csv', index = False)

# Random Forest

# Fit forecaster

regressor_rf = RandomForestRegressor(n_estimators=100, max_samples = 0.4)

forecaster_rf = ForecasterAutoregMultiSeries(
                 regressor          = regressor_rf, 
                 transformer_series = None,
                 lags               = 4,  
                 dropna_from_series = False
             )

%time forecaster_rf.fit(series=series_dict_train_set1, exog=exog_dict_train_set1, suppress_warnings=True)

forecaster_rf

# Feature importances
feat_rf = forecaster_rf.get_feature_importances()

# save file
feat_rf.to_csv('D:/PhD_project/data/results/rf_feat_imp_skt_cv1.csv', index = False)

# Predict - CV1
boot = 1000

%time predictions_rf_cv1 = forecaster_rf.predict_bootstrapping(steps=3, exog=exog_dict_test_set1, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_rf_cv1.keys())

# Create an empty DataFrame
df = pd.DataFrame(columns=['yearmonth', 'id', 'model', 'horizon'] + [f'X_{i}' for i in range(boot)])
df

for i in pred_id:
    
    df_test = predictions_rf_cv1[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'rf'
    
    df_test['horizon'] = 1

    df = pd.concat([df, df_test])

# update forecaster with new data
forecaster_rf.fit(series=series_dict_train_set2, exog=exog_dict_train_set2, suppress_warnings=True)

# Predict - CV2
%time predictions_rf_cv2 = forecaster_rf.predict_bootstrapping(steps=3, exog=exog_dict_test_set2, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_rf_cv2.keys())

for i in pred_id:
    
    df_test = predictions_rf_cv2[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'rf'
    
    df_test['horizon'] = 2

    df = pd.concat([df, df_test])

# Feature importances
feat_rf = forecaster_rf.get_feature_importances()

# save file
feat_rf.to_csv('D:/PhD_project/data/results/rf_feat_imp_skt_cv2.csv', index = False)

# update forecaster with new data
forecaster_rf.fit(series=series_dict_train_set3, exog=exog_dict_train_set3, suppress_warnings=True)

# Predict - CV3
%time predictions_rf_cv3 = forecaster_rf.predict_bootstrapping(steps=3, exog=exog_dict_test_set3, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_rf_cv3.keys())

for i in pred_id:
    
    df_test = predictions_rf_cv3[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'rf'
    
    df_test['horizon'] = 3

    df = pd.concat([df, df_test])

# Feature importances
feat_rf = forecaster_rf.get_feature_importances()

# save file
feat_rf.to_csv('D:/PhD_project/data/results/rf_feat_imp_skt_cv3.csv', index = False)

# save file

df.to_csv('D:/PhD_project/data/predictions/rf_forecasts_skt_cv.csv', index = False)

# MLR

# Fit forecaster

regressor_lr = LinearRegression()

forecaster_lr = ForecasterAutoregMultiSeries(
                 regressor          = regressor_lr, 
                 transformer_series = None,
                 lags               = 4,  
                 dropna_from_series = False
             )

forecaster_lr.fit(series=series_dict_train_set1, exog=exog_dict_train_set1, suppress_warnings=True)

forecaster_lr

# Feature importances
feat_lr = forecaster_lr.get_feature_importances()

# save file
feat_lr.to_csv('D:/PhD_project/data/results/lr_feat_imp_skt_cv1.csv', index = False)

# Predict - CV1

boot = 1000

%time predictions_lr_cv1 = forecaster_lr.predict_bootstrapping(steps=3, exog=exog_dict_test_set1, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_lr_cv1.keys())

# Create an empty DataFrame
df = pd.DataFrame(columns=['yearmonth', 'id', 'model', 'horizon'] + [f'X_{i}' for i in range(boot)])
df

for i in pred_id:
    
    df_test = predictions_lr_cv1[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'mlr'
    
    df_test['horizon'] = 1

    df = pd.concat([df, df_test])

# update forecaster with new data
forecaster_lr.fit(series=series_dict_train_set2, exog=exog_dict_train_set2, suppress_warnings=True)

# Predict - CV2
%time predictions_lr_cv2 = forecaster_lr.predict_bootstrapping(steps=3, exog=exog_dict_test_set2, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_lr_cv2.keys())

for i in pred_id:
    
    df_test = predictions_lr_cv2[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'mlr'
    
    df_test['horizon'] = 2

    df = pd.concat([df, df_test])

# Feature importances
feat_lr = forecaster_lr.get_feature_importances()

# save file
feat_lr.to_csv('D:/PhD_project/data/results/lr_feat_imp_skt_cv2.csv', index = False)

# update forecaster with new data
forecaster_lr.fit(series=series_dict_train_set3, exog=exog_dict_train_set3, suppress_warnings=True)

# Predict - CV3
%time predictions_lr_cv3 = forecaster_lr.predict_bootstrapping(steps=3, exog=exog_dict_test_set3, suppress_warnings=True, n_boot = boot)

pred_id = list(predictions_lr_cv3.keys())

for i in pred_id:
    
    df_test = predictions_lr_cv3[i]
    
    df_test = df_test.reset_index()
    
    df_test.columns = ['yearmonth'] + [f'X_{i}' for i in range(boot)]
    
    df_test['id'] = i
    
    df_test['model'] = 'mlr'
    
    df_test['horizon'] = 3

    df = pd.concat([df, df_test])

# Feature importances
feat_lr = forecaster_lr.get_feature_importances()

# save file
feat_lr.to_csv('D:/PhD_project/data/results/lr_feat_imp_skt_cv3.csv', index = False)

# save file

df.to_csv('D:/PhD_project/data/predictions/lr_forecasts_skt_cv.csv', index = False)
