==> scripts/forecasting_models/cross_validation/04_Nixtla_timeGPT_cv.ipynb <==
!pip install -Uqq nixtla

# Load libraries

from nixtla import NixtlaClient
import pandas as pd
import numpy as np

# initiate Nixtla client

nixtla_client = NixtlaClient(
    api_key = 'nixtla-tok-fQ023pkMcbq0hsxZCUNUyHruTjNHOe07YkG6PIdlUduZaQeyXG18s1e59Fjt2FieAZBWOPxBKTgEeJeD'
)

# key validation

nixtla_client.validate_api_key()

# Load data

from google.colab import files
uploaded = files.upload()

# read data

train_tscv = pd.read_csv("train_tscv (1).csv")
test_tscv = pd.read_csv("test_tscv (1).csv")

# make the yearmonth column as date format

train_tscv['yearmonth'] = pd.to_datetime(train_tscv['yearmonth']) + pd.offsets.MonthEnd(0)
train_tscv = train_tscv.sort_values(by = 'yearmonth')


test_tscv['yearmonth'] = pd.to_datetime(test_tscv['yearmonth']) + pd.offsets.MonthEnd(0)
test_tscv = test_tscv.sort_values(by = 'yearmonth')


train_tscv = train_tscv.rename(columns={
    'yearmonth': 'ds',
    'stock_distributed': 'y'
})

test_tscv = test_tscv.rename(columns={
    'yearmonth': 'ds',
    'stock_distributed': 'y'
})

train_tscv.head()

# define parameters

horizon = 3

# Define prediction intervals
levels = np.arange(0.01, 1.00, 0.01)

# create bounded data to tackle zero demand - this is to avoid negative forecasts

df_transformed = train_tscv.copy()

df_transformed['y'] = np.log(df_transformed['y']+1)

# model fit

timegpt_quantile_fcst_df = nixtla_client.forecast(
    df=df_transformed,
    h=horizon,
    quantiles=levels,
    X_df=test_tscv,
    freq = 'M',
)

# bring back to normal scale

cols_to_transform = [col for col in timegpt_quantile_fcst_df if col not in ['unique_id', 'ds']]

for col in cols_to_transform:
    timegpt_quantile_fcst_df[col] = np.exp(timegpt_quantile_fcst_df[col])-1

timegpt_quantile_fcst_df

# save file

timegpt_quantile_fcst_df.to_csv('timegpt_quantile_cv.csv', index = False)
files.download("timegpt_quantile_cv.csv")
