==> scripts/forecasting_models/cross_validation/06_lag_llama_cv.ipynb <==
!git clone https://github.com/time-series-foundation-models/lag-llama/

cd /content/lag-llama

!pip install -r requirements.txt --quiet # this could take some time # ignore the errors displayed by colab

# download the pretrained model weights from HuggingFace 🤗

!huggingface-cli download time-series-foundation-models/Lag-Llama lag-llama.ckpt --local-dir /content/lag-llama

# import the required packages and the lag llama estimator object

from itertools import islice

from matplotlib import pyplot as plt
import matplotlib.dates as mdates
from tqdm.autonotebook import tqdm

import torch
from gluonts.evaluation import make_evaluation_predictions, Evaluator
from gluonts.dataset.repository.datasets import get_dataset

from gluonts.dataset.pandas import PandasDataset
import pandas as pd

from lag_llama.gluon.estimator import LagLlamaEstimator

# Lag-Llama prediction function

We create a function for Lag-Llama inference that we can reuse for all different types of dataset below. This function returns the predictions for the given prediction horizon. The forecast will be of shape (num_samples, prediction_length), where num_samples is the number of samples sampled from the predicted probability distribution for each timestep.

def get_lag_llama_predictions(dataset, prediction_length, device, context_length=32, use_rope_scaling=False, num_samples=100, nonnegative_pred_samples=True):
    ckpt = torch.load("lag-llama.ckpt", map_location=device) # Uses GPU since in this Colab we use a GPU.
    estimator_args = ckpt["hyper_parameters"]["model_kwargs"]

    rope_scaling_arguments = {
        "type": "linear",
        "factor": max(1.0, (context_length + prediction_length) / estimator_args["context_length"]),
    }

    estimator = LagLlamaEstimator(
        ckpt_path="lag-llama.ckpt",
        prediction_length=prediction_length,
        context_length=context_length, # Lag-Llama was trained with a context length of 32, but can work with any context length

        # estimator args
        input_size=estimator_args["input_size"],
        n_layer=estimator_args["n_layer"],
        n_embd_per_head=estimator_args["n_embd_per_head"],
        n_head=estimator_args["n_head"],
        scaling=estimator_args["scaling"],
        time_feat=estimator_args["time_feat"],
        rope_scaling=rope_scaling_arguments if use_rope_scaling else None,

        nonnegative_pred_samples=True,

        batch_size=8,
        num_parallel_samples=100,
        device=device,
    )

    lightning_module = estimator.create_lightning_module()
    transformation = estimator.create_transformation()
    predictor = estimator.create_predictor(transformation, lightning_module)

    forecast_it, ts_it = make_evaluation_predictions(
        dataset=dataset,
        predictor=predictor,
        num_samples=num_samples
    )
    forecasts = list(forecast_it)
    tss = list(ts_it)

    return forecasts, tss

# Load data

from google.colab import files
uploaded = files.upload()

# read data

stock_df = pd.read_csv("stock_tsb_zero_remove.csv")

# make the yearmonth column as date format

stock_df['yearmonth'] = pd.to_datetime(stock_df['yearmonth']) + pd.offsets.MonthEnd(0)

stock_df = stock_df.sort_values(by = 'yearmonth')

stock_df = stock_df.set_index('yearmonth')

stock_df.head()

# create test and train dataset

series = stock_df[['id', 'stock_distributed']]

series_train = series[series.index <= '2019-9-30']
series_test = series[series.index > '2019-9-30']

# Create the Pandas

series = PandasDataset.from_long_dataframe(series, target="stock_distributed", item_id="id", freq = "1M")
#series_train = PandasDataset.from_long_dataframe(series_train, target="stock_distributed", item_id="id", freq = "1M")
#series_test = PandasDataset.from_long_dataframe(series_test, target="stock_distributed", item_id="id", freq = "1M")

# define parameters

prediction_length = 3
context_length = 32
num_samples = 100
device = torch.device("cuda")

# Get predictions with zero-shot inference

forecasts, tss = get_lag_llama_predictions(series, prediction_length, device, num_samples)

# Create an empty DataFrame to store predicted values

prob_pred = pd.DataFrame(columns=['yearmonth', 'id', 'model'] + [f'X{i}' for i in range(num_samples)])
prob_pred

# extract values

n = len(forecasts)

for m in range(n):

  sample_fc = forecasts[m]
  # Extract data
  item_id = sample_fc.item_id
  samples = sample_fc.samples

  # Create DataFrame
  num_columns = samples.shape[0]
  df = pd.DataFrame(samples.T, columns=[f'X{i}' for i in range(num_columns)])
  # Add item_id and yearmonth columns
  df['id'] = item_id
  df['yearmonth'] = [sample_fc.start_date + i for i in range(samples.shape[1])]
  df['model'] = 'lag_llama'

  # Reorder columns
  df = df[['id', 'yearmonth', 'model'] + [f'X{i}' for i in range(num_columns)]]

  # Creating final df
  prob_pred = pd.concat([prob_pred, df])


prob_pred

evaluator = Evaluator()
agg_metrics, ts_metrics = evaluator(iter(tss), iter(forecasts))
print("CRPS:", agg_metrics['mean_wQuantileLoss'])

# save file

prob_pred.to_csv('lag_llama_pred.csv', index = False)
files.download("lag_llama_pred.csv")
